Exclude strain BeH505764
The M segment for strain BeH505764 (accession PP357049) was grouping near the vary base of the tree and this was screwing up the ability to have a reliable clock. In general you want a divergence tree that looks appropriately ultrametric, ie tips fall where you'd expect them to.

Excluding this sequence, largely resolves the issue with the root-to-tip regression of segment M. I chose to exclude strain BeH505764 for segments L and S for good measure.