Update workflows to use new ingest outputs
Updates the files which ingest uploads and makes the corresponding
changes to the phylogenetic workflow. As metadata (and sequences) now
use "strain" as the unique ID a number of simplifications can be made
to the workflow.

There is one regression: the "accession" column no longer exists and
is thus not exported. We'll fix this in a subsequent commit.