[phylo] export segment specific metadata
This adds back in segment-specific metadata to the Auspice JSON. There
are multiple ways this can be done, each with trade-offs. The approach
employed here leaves the "_{segment}" suffix on the field names.
Alternatively we could remap the metadata file for each `export` call so
that (e.g.) "accession_S" becomes "accession".