[ingest] fix metadata conflicts across segments
Mismatched field values across segments (e.g. segments disagree on the
'date') are now resolved by choosing the most common occurrence with
the intention they are resolved upstream, as implemented here.

This approach was the third implementation. Initially I resolved
disagreements within `group_segments.py` via a provided resolutions
YAML. After discussion with @joverlee521 we decided this could be better
implemented via `augur curate` and the original implementation here did
this _after_ the segment grouping, however this made it impossible to
distinguish disagreements which will be fixed vs those which won't¹

NOTE: Here we use accession as the ID, however using strain name would
be better going forward as it would reduce the duplication needed in the
current format. We can't (currently) do this in oropouche because strain
names are added _after_ the curate chain runs.

¹ <https://github.com/nextstrain/oropouche/pull/18#discussion_r1791090493>